+++ tests/handlers/websocket/test_handle_websocket.py
@@ -1,0 +1,18 @@
+from starlite import WebSocket, create_test_client, websocket
+
+
+def test_handle_websocket():
+    @websocket(path="/")
+    async def simple_websocket_handler(socket: WebSocket) -> None:
+        await socket.accept()
+        data = await socket.receive_json()
+        assert data
+        await socket.send_json({"data": "123"})
+        await socket.close()
+
+    client = create_test_client(route_handlers=simple_websocket_handler)
+
+    with client.websocket_connect("/") as ws:
+        ws.send_json({"data": "123"})
+        data = ws.receive_json()
+        assert data
